requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
pymysql
python-dotenv